\name{Planesmuestra-package}
\alias{Planesmuestra-package}
\alias{Planesmuestra}
\docType{package}
\title{
Acceptance sampling plan according the Dodge Romig, MIL STD105E and MIL STD414 plans.
}
\description{
Use a funcion for each plan and a special one for graphic an OC curve. The plan functions are bassed in the Dodge Romig, MIL STD 105E and MIL STD 414. However, the OC curve is calculated using the binomial trials, after calculating acceptance sampling plan.
}
\details{
\tabular{ll}{
Package: \tab Planesmuestra\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-17\cr
License: \tab GPL\cr
}
}
\author{
Erick Marroquin\cr Maintainer: Erick Marroquin <ericksuhel@gmail.com>
}
\keyword{ package }
